Pics and CRS
I have added some pics that show how the map should look at various
stages of the tutorial. I have also added an 18th century painting of
the River Thames. I also added a step on how to set the CRS. It is odd
that I could not find a England or London CRS, but setting it by layer
worked just fine.